refactor: better testing
