+++ unit_test/main_tests/main_requires_python_test.py
@@ -1,6 +1,5 @@
 import sys
 import textwrap
-from pathlib import Path
 
 import pytest
 from packaging.specifiers import SpecifierSet
@@ -19,11 +18,7 @@ def fake_package_dir(monkeypatch, tmp_path):
 
     local_path.joinpath("setup.py").touch()
 
-    def mock_iterdir(path):
-        return {}
-
     monkeypatch.setattr(sys, 'argv', ['cibuildwheel', str(local_path)])
-    monkeypatch.setattr(Path, 'iterdir', mock_iterdir)
 
     return local_path
